requirements.txt
# Playwright Python BDD Framework Dependencies

# Core Framework
playwright==1.40.0
behave==1.2.6
pytest==7.4.3

# Data Providers
openpyxl==3.1.2
xlrd==2.0.1
pyodbc==5.0.1
pymssql==2.2.11
sqlite3-python==1.0.0

# OCR and Image Processing
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.22.0

# Configuration
configparser==6.0.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Reporting
behave-html-formatter==0.9.10
allure-behave==2.13.2
jinja2==3.1.2
beautifulsoup4==4.12.2

# Email
secure-smtplib==0.1.1
python-email-validator==2.1.0

# Jira Integration
jira==3.5.2
requests==2.31.0

# Utilities
python-dateutil==2.8.2
arrow==1.3.0
faker==20.1.0

# System Monitoring
psutil==5.9.6

# Logging
colorlog==6.8.0
loguru==0.7.2

# Development
black==23.12.1
flake8==6.1.0
pylint==3.0.3